+++ tests/Database/SelectFields/ValidateDiffNodeTests/EpisodeEnum.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rebing\GraphQL\Tests\Database\SelectFields\ValidateDiffNodeTests;
+
+use Rebing\GraphQL\Support\EnumType;
+
+class EpisodeEnum extends EnumType
+{
+    protected $attributes = [
+        'name' => 'Episode',
+        'description' => 'The types of demographic elements',
+        'values' => [
+            'NEWHOPE' => 'NEWHOPE',
+            'EMPIRE' => 'EMPIRE',
+            'JEDI' => 'JEDI',
+        ],
+    ];
+}