+++ tests/Database/SelectFieldsTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Tests\Database;
 
+use Illuminate\Support\Carbon;
 use Rebing\GraphQL\Tests\TestCaseDatabase;
 use Rebing\GraphQL\Tests\Support\Models\Post;
 use Rebing\GraphQL\Tests\Support\Models\Comment;
@@ -18,6 +19,7 @@ use Rebing\GraphQL\Tests\Support\Queries\PostNonNullWithSelectFieldsAndModelQuer
 use Rebing\GraphQL\Tests\Support\Queries\PostsListOfWithSelectFieldsAndModelQuery;
 use Rebing\GraphQL\Tests\Support\Types\PostWithModelAndAliasAndCustomResolverType;
 use Rebing\GraphQL\Tests\Support\Queries\PostWithSelectFieldsAndModelAndAliasQuery;
+use Rebing\GraphQL\Tests\Support\Queries\PostWithSelectFieldsAndModelAndAliasCallbackQuery;
 use Rebing\GraphQL\Tests\Support\Queries\PostsNonNullAndListOfWithSelectFieldsAndModelQuery;
 use Rebing\GraphQL\Tests\Support\Queries\PostWithSelectFieldsAndModelAndAliasAndCustomResolverQuery;
 use Rebing\GraphQL\Tests\Support\Queries\PostsNonNullAndListAndNonNullOfWithSelectFieldsAndModelQuery;
@@ -295,17 +297,26 @@ SQL
             'title' => 'Description of the post',
         ]);
 
-        factory(Comment::class, 2)
+        Carbon::setTestNow('2018-01-01');
+
+        factory(Comment::class)
             ->create([
                 'post_id' => $post->id,
+                'created_at' => new Carbon('2000-01-01'),
+            ]);
+
+        factory(Comment::class)
+            ->create([
+                'post_id' => $post->id,
+                'created_at' => new Carbon('2018-05-05'),
             ]);
 
         $graphql = <<<GRAQPHQL
     {
-      postWithSelectFieldsAndModelAndAlias(id: $post->id) {
+      postWithSelectFieldsAndModelAndAliasCallback(id: $post->id) {
         id
         description
-        commentCount
+        commentsLastMonth
       }
     }
 GRAQPHQL;
@@ -318,7 +329,7 @@ GRAQPHQL;
 
         $this->assertSqlQueries(
             <<<'SQL'
-select "posts"."id", "posts"."title", (SELECT count(*) FROM comments WHERE posts.id = comments.post_id) AS commentCount from "posts" where "posts"."id" = ? limit 1;
+select "posts"."id", "posts"."title", (SELECT count(*) FROM comments WHERE posts.id = comments.post_id AND DATE(created_at) > '2018-01-01 00:00:00') AS commentsLastMonth from "posts" where "posts"."id" = ? limit 1;
 SQL
         );
 
@@ -326,10 +337,10 @@ SQL
 
         $expectedResult = [
             'data' => [
-                'postWithSelectFieldsAndModelAndAlias' => [
+                'postWithSelectFieldsAndModelAndAliasCallback' => [
                     'id' => '1',
                     'description' => 'Description of the post',
-                    'commentCount' => 2,
+                    'commentsLastMonth' => 1,
                 ],
             ],
         ];
@@ -433,6 +444,7 @@ SQL
                 PostWithSelectFieldsAndModelAndAliasQuery::class,
                 PostWithSelectFieldsAndModelQuery::class,
                 PostWithSelectFieldsNoModelQuery::class,
+                PostWithSelectFieldsAndModelAndAliasCallbackQuery::class,
             ],
         ]);