+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 
 ### Fixed
 - Support adding Schema objects directly [\#449 / mfn](https://github.com/rebing/graphql-laravel/pull/449)
+- Input arguments are properly parsed when objects or lists are passed [\#419 / sowork](https://github.com/rebing/graphql-laravel/pull/419)
 
 2019-08-05, 2.0.0
 -----------------